pyproject.toml
[project]
name = "research-service"
version = "0.1.0"
description = "HTTP service for autonomous multi-step web research using PydanticAI and Firecrawl"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-ai[openai,anthropic]==1.56.0",
    "numpy",
    "tavily-python",
    "firecrawl-py",
    "redis[hiredis]",
    "pydantic-settings",
    "httpx",
    "sse-starlette",
    "python-json-logger>=3.0.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "fakeredis",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
